controlled redundant request to server for same purpose especially on slow network

+++ client/src/components/HomePage/RecipeCard.js
@@ -1,3 +1,4 @@
+import { useState } from 'react';
 import { Link } from 'react-router-dom';
 import useRecipeContext from '../../hooks/useRecipeContext';
 
@@ -6,10 +7,15 @@ import customAlert from '../../utils/customAlert';
 import { RECIPE_BASE_URI } from '../../utils/URIs';
 
 const RecipeCard = ({ recipe }) => {
+  const [isRequestPending, setIsRequestPending] = useState(false);
   const { dispatch } = useRecipeContext();
   const { name, description, imageSRC, _id: recipeID } = recipe;
 
   const removeRecipe = async () => {
+    // Check whether a DELETE request is already made or not.
+    if(isRequestPending) return;
+    else setIsRequestPending(true);
+
     const response = await fetch(`${RECIPE_BASE_URI}/${recipeID}`, {
       method: 'DELETE',
       credentials: 'include'
