+++ client/src/components/RecipePage/RecipeUpdateBox.js
@@ -1,3 +1,4 @@
+import { useState } from 'react';
 import { createSearchParams, useNavigate } from 'react-router-dom';
 
 // Utility Stuff.
@@ -5,6 +6,7 @@ import customAlert from '../../utils/customAlert';
 import { RECIPE_BASE_URI } from '../../utils/URIs';
 
 const RecipeUpdateBox = ({ recipeID }) => {
+  const [isRequestPending, setIsRequestPending] = useState(false);
   const navigate = useNavigate();
 
   const navigateBackward = () => {
@@ -12,6 +14,10 @@ const RecipeUpdateBox = ({ recipeID }) => {
   }
 
   const removeRecipe = async () => {
+    // Check whether a DELETE request is already made or not.
+    if(isRequestPending) return;
+    else setIsRequestPending(true);
+
     const response = await fetch(`${RECIPE_BASE_URI}/${recipeID}`, {
       method: 'DELETE',
       credentials: 'include'
